requirements.txt
cffi
git+git://github.com/scrapy/scrapy.git@master
cryptography
scrapyd
beautifulsoup4
pillow
boto
arrow
sqlalchemy
psycopg2
